+++ docs/source/website_config/_specimen_web_config.py
@@ -15,6 +15,7 @@ from typing import List, TYPE_CHECKING
 # Include the following if you want to use it in CELERYBEAT_SCHEDULE
 # from celery.schedules import crontab
 
+from crate_anon.crateweb.anonymise_api.constants import ApiSettingsKeys
 from crate_anon.crateweb.config.constants import ResearchDbInfoKeys as RDIKeys
 
 if TYPE_CHECKING:
@@ -459,14 +460,12 @@ PDF_LETTER_FOOTER_HTML = ""
 # =============================================================================
 # See https://crateanon.readthedocs.io/en/latest/website_config/web_config_file.html  # noqa
 
-CHARITY_URL = "http://www.cpft.nhs.uk/research.htm"
-CHARITY_URL_SHORT = "www.cpft.nhs.uk/research.htm"
-LEAFLET_URL_CPFTRD_CLINRES_SHORT = (
-    "www.cpft.nhs.uk/research.htm > CPFT Research Database"
-)
+CHARITY_URL = "https://www.cpft.nhs.uk/research"
+CHARITY_URL_SHORT = "www.cpft.nhs.uk/research"
+LEAFLET_URL_CPFTRD_CLINRES_SHORT = "www.cpft.nhs.uk/research-database"
 
 ANONYMISE_API = {
-    "HASH_KEY": "aaaa CHANGE THIS! aaaa",
-    "ALLOWLIST_FILENAMES": {},
-    "DENYLIST_FILENAMES": {},
+    ApiSettingsKeys.HASH_KEY: "aaaa CHANGE THIS! aaaa",
+    ApiSettingsKeys.ALLOWLIST_FILENAMES: {},
+    ApiSettingsKeys.DENYLIST_FILENAMES: {},
 }